requirements.txt
fastapi
uvicorn
ultralytics
pillow
opencv-python-headless
